chore: normalize settings, drop old nodes, update deps
BREAKING CHANGE: This drops support for node < 10.

There are some lint failures due to standard pushing for using WhatWG URL
objects instead of url.parse/url.resolve.  However, the code in this lib
does some fancy things with the query/search portions of the parsed url
object, so it'll take a bit of care to make it work properly.