Put default timeout back to zero
Re: https://github.com/npm/npm-registry-fetch/issues/26
Re: https://github.com/npm/cli/issues/1151

The documented default timeout of 30s was not being set in v4, so we
"fixed the glitch" in v4.0.3, causing problems for people trying to
download large packages.

There's no npm v6 way to specify what timeout to use, so not having a
timeout at all seems like a reasonable default for the v4 line, at
least. Let's roll back that change, and document it. (Arguably, fixing
this bug was a breaking change, and we ought to roll it back.)

This effectively reverts 69c297732cfe2d23f0303ab931a7b02364107388, with
documentation of the effective behavior before the change.