Load and send auth based on URI, not registry
This refactors the handling of auth in the new context where npm/cli
will _only_ ever send authorization in such a way that any
authorization-related options are scoped to a given sanitized registry
host and path.

This fixes a troubling situation where a user has intentionally logged
into two separate registries, one which serves package distribution
tarballs, and the other which has packuments pointing to it.

For example, a request to `https://registry.internal/private-thing`
might return this packument:

```json
{
  "name": "private-thing",
  "dist-tags": {
    "latest": "1.0.0"
  },
  "versions": {
    "1.0.0": {
      "name": "private-thing",
      "version": "1.0.0",
      "dist": {
        "tarball": "https://tarballs.internal/private-thing-1.0.0.tgz"
      }
    }
  }
}
```

If the user has properly logged into both `https://registry.internal`
and `https://tarballs.internal` (or at least placed an auth or bearer
token in their `.npmrc` for each), then this should work properly when
they run:

```
npm install private-thing --registry=https://registry.internal
```

However, previously the auth for `https://tarballs.internal` would never
be loaded, because it was being loaded based solely on the _registry_
URI, regardless of the actual URI being fetched, and then only _sent_
when the hostname matches, and so the request for the tarball would
fail.

BREAKING CHANGES:

* The `alwaysAuth` config is no longer relevant.  If we have
  authentication information for a given URI, we send it when making the
  request to that URI.  This was previously being done based on the
  presence of a `scope` parameter or a scoped `spec` option.  However,
  that is not reliable, as users can depend directly on tarball URLs,
  and other parts of the system may have reason to make requests to the
  registry in question without a specific dependency spec in mind (for
  example, to fetch a packument for the purpose of calculating
  metavulnerabilities during an `npm audit` operation.)

* A top level `_auth`, `_authToken`, `username`, `_password`, or
  `password` option is no longer respected if not scoped to a given
  registry URL.  This functionality was removed from npm/cli as of
  version 7, due to the risk of sending credentials to an incorrect host
  by mistake.  The npm cli will automatically scope these top-level
  configs to the default registry option and save them back to the
  `.npmrc` file the first time they are encountered, so users generally
  see no disruption, and this change only affects non-npm users of
  npm-registry-fetch.