fix: consistent use of url.URL
Even though the engines field limits the range of node versions to those
that have a global URL, all but one of the files still used
require('url').URL.  This fixes the one remaining file to also use that.
While older versions of node are still not supported, and not
recommended, this will allow this module to potentially still work there
for the time being.